requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Flask>=2.0.0
flask-cors>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
WTForms>=3.0.0
Werkzeug>=2.0.0
bcrypt>=4.0.0
email-validator>=1.3.0
